remove the new language subtype settings functionality
