+++ app/src/main/java/rkr/simplekeyboard/inputmethod/latin/RichInputMethodManager.java
@@ -18,7 +18,6 @@ package rkr.simplekeyboard.inputmethod.latin;
 
 import android.content.Context;
 import android.content.SharedPreferences;
-import android.content.res.XmlResourceParser;
 import android.inputmethodservice.InputMethodService;
 import android.os.AsyncTask;
 import android.os.Build;
@@ -28,11 +27,6 @@ import android.view.inputmethod.InputMethodInfo;
 import android.view.inputmethod.InputMethodManager;
 import android.view.inputmethod.InputMethodSubtype;
 
-import org.xmlpull.v1.XmlPullParser;
-import org.xmlpull.v1.XmlPullParserException;
-
-import java.io.IOException;
-import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
@@ -55,17 +49,6 @@ import static rkr.simplekeyboard.inputmethod.latin.common.Constants.Subtype.KEYB
 public class RichInputMethodManager {
     private static final String TAG = RichInputMethodManager.class.getSimpleName();
     private static final boolean DEBUG = false;
-
-    private static final String SUBTYPE_META_DATA_NAME = "android.view.im";
-    private static final String XML_NAMESPACE = "http://schemas.android.com/apk/res/android";
-    private static final String TAG_SUBTYPE = "subtype";
-    private static final String TAG_ICON = "icon";
-    private static final String TAG_LABEL = "label";
-    private static final String TAG_SUBTYPE_ID = "subtypeId";
-    private static final String TAG_LOCALE = "imeSubtypeLocale";
-    private static final String TAG_MODE = "imeSubtypeMode";
-    private static final String TAG_EXTRA_VALUE = "imeSubtypeExtraValue";
-    private static final String TAG_ASCII_CAPABLE = "isAsciiCapable";
 
     private RichInputMethodManager() {
         // This utility class is not publicly instantiable.
@@ -495,74 +478,6 @@ public class RichInputMethodManager {
             final boolean allowsImplicitlySelectedSubtypes) {
         return mInputMethodInfoCache.getEnabledInputMethodSubtypeList(
                 imi, allowsImplicitlySelectedSubtypes);
-    }
-
-    /**
-     * Get the list of subtypes that are always available for this IME. This returns the list of
-     * subtypes from method.xml.
-     * @return the list of default subtypes.
-     */
-    public List<InputMethodSubtype> getDefaultSubtypesOfThisIme() {
-        final InputMethodInfo imi = getInputMethodInfoOfThisIme();
-        final int inputMethodResId = imi.getServiceInfo().metaData.getInt(SUBTYPE_META_DATA_NAME);
-        final XmlResourceParser parser = mContext.getResources().getXml(inputMethodResId);
-        final List<InputMethodSubtype> defaultSubtypes = new ArrayList<>();
-        try {
-            while (parser.getEventType() != XmlResourceParser.END_DOCUMENT) {
-                final int event = parser.next();
-                if (event == XmlPullParser.START_TAG) {
-                    final String tag = parser.getName();
-                    if (TAG_SUBTYPE.equals(tag)) {
-                        defaultSubtypes.add(parseSubtype(parser));
-                    }
-                }
-            }
-        } catch (XmlPullParserException | IOException e) {
-            e.printStackTrace();
-            Log.e(TAG, "Failed to parse default subtypes");
-            return new ArrayList<>();
-        }
-        return defaultSubtypes;
-    }
-
-    /**
-     * Build a subtype from an xml tag.
-     * @param parser the XML parser.
-     * @return the subtype that the xml defines.
-     */
-    private static InputMethodSubtype parseSubtype(final XmlResourceParser parser) {
-        final InputMethodSubtype.InputMethodSubtypeBuilder builder =
-                new InputMethodSubtype.InputMethodSubtypeBuilder();
-
-        final int icon = parser.getAttributeResourceValue(XML_NAMESPACE, TAG_ICON, -1);
-        if (icon != -1) {
-            builder.setSubtypeIconResId(icon);
-        }
-
-        final int label = parser.getAttributeResourceValue(XML_NAMESPACE, TAG_LABEL, -1);
-        if (label != -1) {
-            builder.setSubtypeNameResId(label);
-        }
-
-        final int id = parser.getAttributeIntValue(XML_NAMESPACE, TAG_SUBTYPE_ID, -1);
-        if (id != -1) {
-            builder.setSubtypeId(id);
-        }
-
-        final String locale = parser.getAttributeValue(XML_NAMESPACE, TAG_LOCALE);
-        builder.setSubtypeLocale(locale);
-
-        final String mode = parser.getAttributeValue(XML_NAMESPACE, TAG_MODE);
-        builder.setSubtypeMode(mode);
-
-        final String extraValue = parser.getAttributeValue(XML_NAMESPACE, TAG_EXTRA_VALUE);
-        builder.setSubtypeExtraValue(extraValue);
-
-        final boolean asciiCapable = parser.getAttributeBooleanValue(XML_NAMESPACE,
-                TAG_ASCII_CAPABLE, false);
-        builder.setIsAsciiCapable(asciiCapable);
-
-        return builder.build();
     }
 
     public void refreshSubtypeCaches() {
